# backend/swagger/doc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Farmako Coupon API",
    "description": "API for managing and validating pharmacy coupons",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/coupons": {
      "get": {
        "summary": "List all coupons",
        "responses": {
          "200": {
            "description": "A list of coupons",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Coupon"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new coupon",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Coupon"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Coupon created successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/coupons/{id}": {
      "get": {
        "summary": "Get a specific coupon",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A coupon",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Coupon"
                }
              }
            }
          },
          "404": {
            "description": "Coupon not found"
          }
        }
      },
      "put": {
        "summary": "Update a coupon",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Coupon"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Coupon updated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Coupon not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a coupon",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Coupon deleted"
          },
          "404": {
            "description": "Coupon not found"
          }
        }
      }
    },
    "/coupons/validate": {
      "post": {
        "summary": "Validate a coupon",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Coupon validation result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/coupons/applicable": {
      "get": {
        "summary": "Get applicable coupons for a cart",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicableCouponsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of applicable coupons",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicableCouponsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/coupons/apply": {
      "post": {
        "summary": "Apply a coupon to an order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Coupon application result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Coupon": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "coupon_code": {
            "type": "string"
          },
          "expiry_date": {
            "type": "string",
            "format": "date-time"
          },
          "usage_type": {
            "type": "string",
            "enum": ["one_time", "multi_use", "time_based"]
          },
          "applicable_medicine_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "applicable_categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "min_order_value": {
            "type": "number",
            "format": "float"
          },
          "valid_time_window_start": {
            "type": "string",
            "format": "date-time"
          },
          "valid_time_window_end": {
            "type": "string",
            "format": "date-time"
          },
          "terms_and_conditions": {
            "type": "string"
          },
          "discount_type": {
            "type": "string",
            "enum": ["percentage", "fixed"]
          },
          "discount_value": {
            "type": "number",
            "format": "float"
          },
          "max_usage_per_user": {
            "type": "integer"
          }
        },
        "required": ["coupon_code", "expiry_date", "discount_type", "discount_value"]
      },
      "CartItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "float"
          },
          "quantity": {
            "type": "integer"
          }
        },
        "required": ["id", "price", "quantity"]
      },
      "ValidationRequest": {
        "type": "object",
        "properties": {
          "coupon_code": {
            "type": "string"
          },
          "cart_items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItem"
            }
          },
          "order_total": {
            "type": "number",
            "format": "float"
          },
          "user_id": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["coupon_code", "cart_items", "order_total"]
      },
      "Discount": {
        "type": "object",
        "properties": {
          "items_discount": {
            "type": "number",
            "format": "float"
          },
          "charges_discount": {
            "type": "number",
            "format": "float"
          },
          "total_discount": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "ValidationResponse": {
        "type": "object",
        "properties": {
          "is_valid": {
            "type": "boolean"
          },
          "discount": {
            "$ref": "#/components/schemas/Discount"
          },
          "message": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": ["is_valid"]
      },
      "ApplicableCouponsRequest": {
        "type": "object",
        "properties": {
          "cart_items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItem"
            }
          },
          "order_total": {
            "type": "number",
            "format": "float"
          },
          "user_id": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["cart_items", "order_total"]
      },
      "ApplicableCoupon": {
        "type": "object",
        "properties": {
          "coupon_code": {
            "type": "string"
          },
          "discount_value": {
            "type": "number",
            "format": "float"
          },
          "discount_type": {
            "type": "string",
            "enum": ["percentage", "fixed"]
          },
          "expiry_date": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          }
        },
        "required": ["coupon_code", "discount_value", "discount_type", "expiry_date"]
      },
      "ApplicableCouponsResponse": {
        "type": "object",
        "properties": {
          "applicable_coupons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApplicableCoupon"
            }
          }
        }
      }
    }
  }
} 